Select position by time function had int instead of double parameter which was causing small divisions to not be selectable.

+++ xLights/sequencer/EffectLayer.cpp
@@ -420,7 +420,7 @@ void EffectLayer::SelectEffectsInPositionRange(int startX,int endX)
     }
 }
 
-void EffectLayer::SelectEffectsInTimeRange(double startTime,int endTime)
+void EffectLayer::SelectEffectsInTimeRange(double startTime,double endTime)
 {
     for(int i=0;i<mEffects.size();i++)
     {
